+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 * build: deploy dev Android apk to Firebase
 * build: introduce golden tests
 * build: collect tests coverage
+* feat: introduce exercise widget